Delay scope resolution for const, unless global is specified
Expand (const (= x e)) into

  (block
    (= (ssavalue n) e)
    (const x (ssavalue n)
    (latestworld))

so that resolve-scopes (which is modified to treat (const x y) as an assignment
form) can determine if x really refers to a global.  When we linearize, we will
throw an error if x has not been resolved to a globalref, or if the (const x e)
form is inside of a function.

At the top level, these should succeed:

    const x = 1
    begin
        const x = 1
    end
    let
        const global x = 1
    end

Each of the following should fail:

    # ERROR: syntax: unsupported `const` declaration on local variable around REPL[1]:2
    let
        const x = 1
    end

    # ERROR: syntax: unsupported `const` declaration on local variable around REPL[1]:2
    function f()
        const x = 1
    end

    # ERROR: syntax: `global const` declaration not allowed inside function around REPL[1]:2
    function f()
        global const x = 1
    end